+++ src/fs.py
@@ -10,7 +10,24 @@ def shouldOverwrite(Overwrite, Sure, Path):
 	if Sure:
 		return True
 	return question(f"Are you sure you want to overwrite {Path}")
-def createFolder(Dirs, Overwrite=False, Sure=False):
+def updateTime(Path, AccessTime=None, ModifiedTime=None):
+	Time = []
+	Stat = os.stat(Path)
+	if AccessTime is not None and isinstance(AccessTime, (int, float)):
+		Time.append(AccessTime)
+	else:
+		if AccessTime is not None:
+			log(f"Warning: Access time is invalid: {AccessTime}", InfoLevel.normal, sys.stderr)
+		Time.append(Stat.st_atime)
+	if ModifiedTime is not None and isinstance(ModifiedTime, (int, float)):
+		Time.append(ModifiedTime)
+	else:
+		if ModifiedTime is not None:
+			log(f"Warning: Modified time is invalid: {ModifiedTime}", InfoLevel.normal, sys.stderr)
+		Time.append(Stat.st_mtime)
+	os.utime(Path, tuple(Time))
+	log(f"Updated access time of {Path} to {Time[0]} and modified time of {Path} to {Time[1]}", InfoLevel.verbose)
+def createFolder(Dirs, Overwrite=False, Sure=False, ChangeTimestamp=True, AccessTime=None, ModifiedTime=None):
 	if not isinstance(Dirs, list):
 		log(f"{str(Dirs)} is invalid.", InfoLevel.quiet, sys.stderr)
 		return None
@@ -24,6 +41,8 @@ def createFolder(Dirs, Overwrite=False, Sure=False):
 				log(f"{Dir} is not a directory.", InfoLevel.quiet, sys.stderr)
 				continue
 			if not shouldOverwrite(Overwrite, Sure, Dir):
+				if ChangeTimestamp:
+					updateTime(Dir, AccessTime, ModifiedTime)
 				continue
 			try:
 				shutil.rmtree(Dir)
@@ -37,13 +56,15 @@ def createFolder(Dirs, Overwrite=False, Sure=False):
 		try:
 			os.makedirs(Dir, exist_ok=True)
 			log(f"Created {Dir}", InfoLevel.verbose)
+			if ChangeTimestamp:
+				updateTime(Dir, AccessTime, ModifiedTime)
 		except PermissionError:
 			log(f"Permission denied: {Dir}", InfoLevel.quiet, sys.stderr)
 			continue
 		except OSError as Error:
 			log(f"Cannot create directory {Dir}.\n{Error}", InfoLevel.quiet, sys.stderr)
 			continue
-def createFile(Files, Byte=False, Encoding="utf-8", Overwrite=False, Sure=False):
+def createFile(Files, Byte=False, Encoding="utf-8", Overwrite=False, Sure=False, ChangeTimestamp=True, AccessTime=None, ModifiedTime=None):
 	if not isinstance(Files, list):
 		log(f"{str(Files)} is invalid.", InfoLevel.quiet, sys.stderr)
 		return None
@@ -58,6 +79,8 @@ def createFile(Files, Byte=False, Encoding="utf-8", Overwrite=False, Sure=False)
 				log(f"{File} is a directory.", InfoLevel.quiet, sys.stderr)
 				continue
 			if not shouldOverwrite(Overwrite, Sure, File):
+				if ChangeTimestamp:
+					updateTime(File, AccessTime, ModifiedTime)
 				continue
 			Mode = "wb" if Byte else "w"
 		else:
@@ -76,6 +99,8 @@ def createFile(Files, Byte=False, Encoding="utf-8", Overwrite=False, Sure=False)
 				log(f"Overwritten {File}", InfoLevel.verbose)
 			else:
 				log(f"Created {File}", InfoLevel.verbose)
+			if ChangeTimestamp:
+				updateTime(File, AccessTime, ModifiedTime)
 		except PermissionError:
 			log(f"Permission denied: {File}", InfoLevel.quiet, sys.stderr)
 			continue