+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## New version 1.3
+- Added -a; --access-time to update the access time of an existing file or folder or after it has been created.
+- Added -a; ----modified-time to update the modified time of an existing file or folder or after it has been created.
+- Added -c; --no-change-timestamp to not update access time and modified time after creating a new file or folder or when the file or folder already exists.
+- Improved help output.
+- Some internal improvements and bug fixes.
+
+---
+
 ## New version 1.2
 - Some internal improvements and bug fixes.